# lib/api-zod/src/generated/types/bridgeRoute.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * EXIRA Protocol API - 7-pillar behavioral truth services
 * OpenAPI spec version: 1.0.0
 */
import type { BridgeRouteRouteType } from './bridgeRouteRouteType';

export interface BridgeRoute {
  routeId: string;
  routeType: BridgeRouteRouteType;
  sourceChain: number;
  targetChain: number;
  btcpScore: number;
  estimatedGas: string;
  estimatedTime: number;
  nlScore: number;
  beoContruinity: number;
  ccCoherence: number;
}