# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * EXIRA Protocol API - 7-pillar behavioral truth services
 * OpenAPI spec version: 1.0.0
 */

export interface HealthStatus {
  status: string;
}