# lib/api-zod/src/generated/types/signalEvent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * EXIRA Protocol API - 7-pillar behavioral truth services
 * OpenAPI spec version: 1.0.0
 */

export interface SignalEvent {
  signalId: string;
  signalType: string;
  entityId: string;
  coherence: number;
  threshold: number;
  margin?: number;
  /** @nullable */
  direction?: string | null;
  timestamp: string;
}